change return type of method getWorlDim

+++ src/main/java/it/unibo/game/app/api/Model.java
@@ -29,6 +29,6 @@ public interface Model{
 
     Double getRow(Double x);
 
-    Pair<Integer,Integer> getWorldDim();
+    Pair<Double,Double> getWorldDim();
 
 }
